+++ pages/_app.jsx
@@ -1,6 +1,7 @@
 // pages/_app.js [Server Side]
 
 import "@/styles/globals.css";
+import "@/styles/prism-vsc-dark-plus.css";
 
 export default function MyApp({ Component, pageProps }) {
   // Use the layout defined at the page level, if available