+++ pages/blog/[slug].jsx
@@ -1,11 +1,11 @@
 // pages/blog/[slug]/index.js [??? Side]
 
 import { useRouter } from "next/router";
-import BlogLayout from "./../../(layout)/BlogLayout";
-import path from "path";
-import { truncateFileExtensions } from "./../../lib/helper";
-
 import dynamic from "next/dynamic";
+import path from "path";
+
+import BlogLayout from "./../(layout)/BlogLayout";
+import { truncateFileExtensions } from "../../lib/helper";
 
 export default function Page() {
   const router = useRouter();
@@ -19,23 +19,7 @@ export default function Page() {
   );
 }
 
-export async function getStaticProps() {
-  const fs = require("fs");
-  const directoryPath = path.join(process.cwd(), "data/posts");
-  const allFiles = fs.readdirSync(directoryPath, "utf-8");
-  const blogFiles = allFiles.filter((file) => {
-    return file.endsWith(".md") || file.endsWith(".mdx");
-  });
-  const data = truncateFileExtensions(blogFiles);
-  console.log("[slug]/getstaticprops", data);
-  return {
-    // Passed to the page component as props
-    props: {
-      blog: data,
-    },
-  };
-}
-
+// Called First
 export function getStaticPaths() {
   const fs = require("fs");
   const directoryPath = path.join(process.cwd(), "data/posts");
@@ -44,10 +28,28 @@ export function getStaticPaths() {
     return file.endsWith(".md") || file.endsWith(".mdx");
   });
   const data = truncateFileExtensions(blogFiles);
+
   const paths = data.map((slug) => `/blog/${slug}`);
 
   return {
     paths,
     fallback: false,
+  };
+}
+
+// Called Second
+export async function getStaticProps() {
+  const fs = require("fs");
+  const directoryPath = path.join(process.cwd(), "data/posts");
+  const allFiles = fs.readdirSync(directoryPath, "utf-8");
+  const blogFiles = allFiles.filter((file) => {
+    return file.endsWith(".md") || file.endsWith(".mdx");
+  });
+  const data = truncateFileExtensions(blogFiles);
+  return {
+    // Passed to the page component as props
+    props: {
+      blog: data,
+    },
   };
 }